fix: address Trivy normalization review
