+++ packages/shared/src/types/sast-normalization.ts
@@ -6,6 +6,18 @@ import type {
   SastFindingProvenance,
   SastScanLane
 } from './sast-runtime';
+import {
+  hasExactKeys,
+  hasUnsafeControl,
+  isAllowedString,
+  isBoundedIdentifier,
+  isBoundedReference,
+  isCommitSha,
+  isNormalizationScopeValid,
+  isRecord,
+  isSha256Digest,
+  utf8Length
+} from './sast-normalization-validation';
 
 export const OPENGREP_SARIF_NORMALIZER_VERSION =
   'opengrep-sarif-normalizer-v1' as const;
@@ -1101,22 +1113,6 @@ function isTrivyFindingDetailsValid(
   );
 }
 
-function isNormalizationScopeValid(
-  value: unknown
-): value is SastArtifactDispositionScope {
-  return (
-    isRecord(value) &&
-    hasExactKeys(value, [
-      'tenantId',
-      'repositoryBindingId',
-      'scanRequestId',
-      'attemptId',
-      'scannerRunId'
-    ]) &&
-    Object.values(value).every(isBoundedReference)
-  );
-}
-
 function isOrderedIdentifierArray(
   value: unknown,
   pattern: RegExp,
@@ -1176,17 +1172,6 @@ function isSafeNormalizedPath(value: unknown): value is string {
   );
 }
 
-function isBoundedReference(value: unknown): value is string {
-  return isBoundedIdentifier(value, 2048, false);
-}
-
-function isAllowedString(
-  value: unknown,
-  allowed: readonly string[]
-): value is string {
-  return typeof value === 'string' && allowed.includes(value);
-}
-
 function isOptionalBoundedText(
   value: unknown,
   maximumBytes: number,
@@ -1213,38 +1198,6 @@ function isBoundedPlainText(
   );
 }
 
-function isBoundedIdentifier(
-  value: unknown,
-  maximumBytes: number,
-  allowEmpty: boolean
-): value is string {
-  return (
-    typeof value === 'string' &&
-    (allowEmpty || value.length > 0) &&
-    value === value.normalize('NFC') &&
-    value === value.trim() &&
-    utf8Length(value) <= maximumBytes &&
-    !hasUnsafeControl(value, false)
-  );
-}
-
-function hasUnsafeControl(
-  value: string,
-  allowNewlines: boolean
-): boolean {
-  return [...value].some((character) => {
-    const codePoint = character.codePointAt(0) ?? 0;
-    if (allowNewlines && [0x09, 0x0a].includes(codePoint)) {
-      return false;
-    }
-    return (
-      codePoint <= 0x1f ||
-      (codePoint >= 0x7f && codePoint <= 0x9f) ||
-      (codePoint >= 0xd800 && codePoint <= 0xdfff)
-    );
-  });
-}
-
 function isOptionalPositiveCoordinate(value: unknown): boolean {
   return value === undefined || isPositiveCoordinate(value);
 }
@@ -1258,41 +1211,8 @@ function isPositiveCoordinate(value: unknown): value is number {
   );
 }
 
-function isCommitSha(value: unknown): value is string {
-  return (
-    typeof value === 'string' &&
-    /^(?:[a-f0-9]{40}|[a-f0-9]{64})$/u.test(value)
-  );
-}
-
-function isSha256Digest(value: unknown): value is `sha256:${string}` {
-  return (
-    typeof value === 'string' &&
-    /^sha256:[a-f0-9]{64}$/u.test(value)
-  );
-}
-
-function utf8Length(value: string): number {
-  return new TextEncoder().encode(value).byteLength;
-}
-
 function compareCodeUnitStrings(left: string, right: string): number {
   return left < right ? -1 : left > right ? 1 : 0;
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return !!value && typeof value === 'object' && !Array.isArray(value);
-}
-
-function hasExactKeys(
-  value: Record<string, unknown>,
-  expected: readonly string[]
-): boolean {
-  const actual = Object.keys(value);
-  return (
-    actual.length === expected.length &&
-    actual.every((key) => expected.includes(key))
-  );
 }
 
 function hasOnlyKeys(
