+++ specs/006-production-sast-runtime-design/data-model.md
@@ -371,7 +371,9 @@ An in-memory, non-durable T032/T033 handoff from one accepted scanner artifact.
   as a non-authoritative hint and namespaced opaque Trivy structural hashes that never contain
   detected secret values. Trivy secret/IaC identity preimages exclude line/column coordinates;
   coordinates remain occurrence metadata, while a deterministic producer-order ordinal
-  distinguishes multiple findings for the same semantic rule and canonical target
+  distinguishes multiple findings for the same semantic rule and canonical target. Dependency
+  identity includes its canonical package/result target so separate monorepo manifests do not
+  collide
 - capability-discriminated Trivy details: package/advisory state for
   `DEPENDENCY_VULNERABILITY`, category plus explicit payload-discard booleans for
   `SECRET_DETECTION`, or check type/AVD ID/failing result for `IAC_MISCONFIGURATION`
